mmc:core:Disable CRC checking
A huge 30%(according to the comment) gains for IO throughput, though you sacrifice file integrity.

Also allow it to be configurable via its parameter.

note: If you are to copy/use some part of the commit kindly use the original comment/content as such to show respect either ways I'd be obliged to do the same.

Signed-off-by: zeroblade1984 <zeroblade1984@gmail.com>